feat: refactor dependency injection components for modularity
Significantly refactored `AssemblyScanner.cs` by removing private fields and methods, replacing them with new classes: `AssemblySelector`, `TypeSelector`, and `ServiceRegister`. This enhances code organization and readability.

Moved `ExcludeFromRegistrationAttribute.cs` and `AutoRegisterAttribute.cs` to the new namespace `OnForkHub.Core.Attributes` for better attribute organization.

Introduced `AssemblySelector.cs`, `ServiceRegister.cs`, and `TypeSelector.cs` to encapsulate assembly management, service registration logic, and type selection, respectively.

These changes promote a more modular architecture, improving maintainability and separation of concerns within the dependency injection framework.
