+++ src/Core/OnForkHub.Core/Attributes/ExcludeFromRegistrationAttribute.cs
@@ -1,4 +1,4 @@
-namespace OnForkHub.Application.DependencyInjection;
+namespace OnForkHub.Core.Attributes;
 
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
 public class ExcludeFromRegistrationAttribute : Attribute { }
